feat: Move four CompleteMe methods to private
Moved the methods for checking if an input is a string, array, or either
and the method for converting to an array to private since they are only
used locally inside of the class. Also had to delete the tests for those
since you can't test private methods, so my tests are back down to 60.
Need to think of some more tests.